requirements.txt
asgiref==3.5.1
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
djlint==1.0.2
importlib-metadata==4.11.3
Markdown==3.3.7
pathspec==0.9.0
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.1
zipp==3.8.0
